+++ types/mp4box.d.ts
@@ -92,15 +92,15 @@ declare module '@webav/mp4box.js' {
   export interface SampleOpts {
     duration: number;
     dts?: number;
-    cts?: number;
+    cts: number;
     sample_description_index?: number;
     is_sync: boolean;
-    description?: MP4ABoxParser | AVC1BoxParser;
+    description?: MP4ABoxParser | AVC1BoxParser | HVCBoxParser;
   }
 
   export interface MP4Sample {
     track_id: number;
-    description: MP4ABoxParser | AVC1BoxParser;
+    description: MP4ABoxParser | AVC1BoxParser | HVCBoxParser;
     is_rap: boolean;
     is_sync: boolean;
     timescale: number;
@@ -220,7 +220,7 @@ declare module '@webav/mp4box.js' {
   }
 
   export interface HVCBoxParser extends BoxParser {
-    type: 'hvc';
+    type: 'hvc1';
     boxes: HVCCBox[];
     hvcC: HVCCBox;
     compressorname: string;